refactor: 날짜 포맷 기능을 util 클래스로 분리하여 적용
