+++ src/main/java/attendance/model/AttendanceHistory.java
@@ -3,13 +3,11 @@ package attendance.model;
 import static attendance.error.ErrorMessage.ERROR_ATTENDANCE_DETAIL_NOT_FOUND;
 
 import java.time.LocalDate;
-import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Stream;
 
 public class AttendanceHistory {
-    private final DateTimeFormatter formatter = DateTimeFormatter.ofPattern("MM월 dd일 EEEE은 등교일이 아닙니다.");
     private final List<AttendanceDetail> attendanceHistory = new ArrayList<>();
 
     public void addAttendanceDetail(AttendanceDetail attendanceDetail) {
