+++ src/main/java/attendance/model/CustomLocalDateTime.java
@@ -5,10 +5,8 @@ import java.time.Clock;
 import java.time.DayOfWeek;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
-import java.time.LocalTime;
 import java.time.ZoneId;
 import java.time.ZoneOffset;
-import java.time.format.DateTimeFormatter;
 import java.util.Set;
 
 public class CustomLocalDateTime {
@@ -18,7 +16,6 @@ public class CustomLocalDateTime {
             ZoneId.of("UTC")
     );
     private static final Set<LocalDate> holidayDate = Set.of(LocalDate.of(2024, 12, 25));
-    private static final DateTimeFormatter MODIFY_ATTENDANCE_TIME_FORMATTER = DateTimeFormatter.ofPattern("HH:mm");
 
     public static LocalDateTime now() {
         return LocalDateTime.now(clock);
@@ -26,10 +23,6 @@ public class CustomLocalDateTime {
 
     public static LocalDate nowDate() {
         return now().toLocalDate();
-    }
-
-    public static LocalTime parseTime(String time) {
-        return LocalTime.parse(time, MODIFY_ATTENDANCE_TIME_FORMATTER);
     }
 
     public static LocalDate parseDate(String day) {
