+++ src/test/java/attendance/model/AttendanceDetailTest.java
@@ -1,17 +1,16 @@
 package attendance.model;
 
+import static attendance.util.DateFormatUtil.NOT_ATTENDABLE_FORMATTER;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.LocalTime;
-import java.time.format.DateTimeFormatter;
 import org.assertj.core.api.SoftAssertions;
 import org.junit.jupiter.api.Test;
 
 public class AttendanceDetailTest {
-    private final DateTimeFormatter formatter = DateTimeFormatter.ofPattern("MM월 dd일 EEEE은 등교일이 아닙니다.");
 
     @Test
     void 등교일이_아닐때_출석상세_생성시_예외가_발생한다() {
@@ -21,7 +20,7 @@ public class AttendanceDetailTest {
         //when & then
         assertThatThrownBy(() -> new AttendanceDetail(attendanceDateTime))
                 .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage(attendanceDateTime.format(formatter));
+                .hasMessage(attendanceDateTime.format(NOT_ATTENDABLE_FORMATTER));
     }
 
     @Test
